// src/types/poolsFilter.type.ts
export type PoolsFilter = {
    minMillipercent: number;
    maxMillipercent: number;
    minOverdueMillipercent: number;
    maxOverdueMillipercent: number;
    minDuration: number;
    maxDuration: number;
    minFreePeriod: number;
    maxFreePeriod: number;
    minDuty: number;
    maxDuty: number;
    minExpired: number;
    maxExpired: number;
    minPrice: number;
    maxPrice: number
};